Moved classes "_Animation" and "Layers" into timeline.js
